chore: refactor the profiler into scenarios.
